This checkin doubles build times. Sorry.
We now create debug versions of all libraries along with release
(optimize). This makes it much easier for me to debug on different
OSes.

I'm not too worried about this, as I envision eventually the jar files
being in a Maven repository, and 99% of folks never worrying about
building. But for those unfortunate few, yes, your build time just
doubled.