do jarsigning of artifacts with a fake keystore.
This means that people who build this themselves with be signed with
the self-signed development.keystore, which if running as an applet will
display "untrusted" to the user. To change, when running a MVN build,
pass in -Djarsigner.keystore and other jarsigner.* properties to tell
the build system the actual keystore to use.

Note to self: when doing releases, use the actual keystore.