add build and deploy profiles, and auto-enable build.
To build (and install locally), this will work:

    mvn install

However, to deploy you need to do:

    mvn -Pdeploy deploy

Finally, sometimes you want to test a redeploy without a full build. To
do that, do:

    mvn -P!build,deploy

This works by having the build profile run whenver a file is not present
(and that file is never generated, so it always runs). This 'maven hack'
allows us to turn off some projects from always rebuilding. Use with
caution.