move exhaustive tests to their own package.
Alas, with the decision to separate the integration tests into a
different jar, and signing jars, we run into a java exception if I use
io.humble.ferry as the package name for these tests.

The unfortunately side effect of this is I have to make a
package-private member of JNIMemoryManager public.