Initialize `mSymbols` to NULL if symbols of referenced `HumbleStackTrace` is empty
If `mSymbols` field of passed `HumbleStackTrace` instance is empty, `mSymbols` field of target `HumbleStackTrace` instance is not set to NULL in its constructor. In this case, if `HumbleStackTrace` instance is allocated on stack (for example by just declaring it like `HumbleStackTrace hst(targetHst);` as variable), its `mSymbols` field is not set to NULL so it might have a random content. Therefore while it is being destroyed, in its destructor, it might try to free an invalid memory location by `free(mSymbols)` call.