fix #122 jar files exported from eclipse is too huge.
My fix is to lean-into maven, and now dragging a dependency on
`humble-video-all` will figure out what JVM you're running under
and only bring in the native code for the platform you're on.

If you want other profiles, run maven with -P... with the name of
the profile for the OS you want.